pyproject.toml
[tool.poetry]
name = "certificates-backend"
version = "0.1.0"
description = ""
authors = ["Arpan Pandey <arpan@hackersreboot.tech>"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.18.3"
fastapi = "^0.81.0"
Pillow = "^9.2.0"
SQLAlchemy = "^1.4.41"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
PyMySQL = "^1.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
